assign GPIO2 to Cortex-M4 exclusively
It turns out that Linux actually handels interrupts and clears the
interrupt stats register as well as mask the IRQ! This is problematic
and can easily lead to race conditions. Allocate the whole GPIO2 bank
to the Cortex-M4, and enforce this using RDC (Resource Domain
Controller). This might crash Linux in case the kernel tries to access
the bank! Make sure to disable/remove GPIO2 from the device tree.